backend/requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==4.2.2
cattrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.1
exceptiongroup==1.0.0rc9
google-api-core==2.10.0
google-api-python-client==2.60.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
greenlet==1.1.3
httplib2==0.20.4
idna==3.3
protobuf==4.21.5
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dotenv==0.21.0
python-telegram-bot==13.14
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-cache==0.9.6
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
tornado==6.1
tzdata==2022.2
tzlocal==4.2
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==1.26.12
